pyproject.toml
[project]
name = "chameleon-quant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.10.*"
dependencies = [
    "bokeh>=3.6.3",

    "bokeh-sampledata>=2024.2",
    "ipykernel>=6.29.5",
    "numba>=0.60.0",
    "numpy==1.26.4",
    "setuptools>=75.6.0",
    "optunahub>=0.1.0",
    "optuna-dashboard>=0.17.0",
    "typer>=0.15.1",
    "niquests>=3.12.1",
    "fastapi>=0.115.6",
    "uvicorn>=0.34.0",
    "pandas-ta>=0.3.14b0",
    "ccxt>=4.4.50",
    "tqdm>=4.67.1",
    "ipywidgets>=8.1.5",
    "pathos>=0.3.3",
    "anywidget>=0.9.13",
]


[tool.pdm]

[tool.uv.sources]

